tab layout fixed, made dynamic buttons
buttons now have a prop that specifies if you can control the width or have it wrap to the contents which makes making buttons in the future a lot easier. I also fixed the tab layout to be a little lower and removed the headers and made the tab bars transparent.