categories subsection in item!
added a new parameter in itemcontext that is one of these catagories (for now) and you can slide a bar to select them. right now im still working on implementing the actuall adding function to the list but its looking good so far, i just have to edit the list.tsx to change the items array and add the param.