fixed category selection, added home shortcut
fixed the category selection to have the dropdown appear for android and style it separately from iOS since they use different things but now it works really good. also added a route from the main page to the new item page for pantry and added a cooldown for it so it doesn't instantly navigate and the user can understand where the button is taking them. so far so good :)