amount description for item!
added a param that describes how much of an item you have. this makes it so that when people use an item they can delete or add more easily, which is the next feature. also added an example viewbar at the top that shows how the items are made up. so far so good :)