trying to make 1st build, fixing layouts
tried making the 1st build and realized i was wasting a lot of time, instead i decided to put that off until i get 50 hrs so i am still on track. also working on fixing up the layout and padding for android and ios. there are also multiple styling things i need to fix for android. so far so good tho :)