separate editable units!
instead of making the units a seperate category like i should have, i just merged it into the string which was not a good choice at that time, but i guess i had other things to worry about. now i seperated those things and added a new param that only gets put together to show how much of the count you have next to the unit. this will then be used to easily remove and add 1 of an item based on the unit instead of having to rip apart a string. this went pretty smoothly and i felt more confident with all of my code this time. so far so good :)