web style fixes
changed styling in the web to accurately reflect how it looks on mobile (keep in mind it is best viewed on mobile website because the app is designed to be for mobile and i made the web design to be viewed on a mobile browser. ) i also ended up changing the normal styles on the way  so i guess thats also good. so far so good :)