implemented cloud item storage
i cant even explain the pain of 4 hours i went through to get this to work. basically when creating a new item, it adds it to a doc that then is subscibed to by the list via searching the data for the doc. the only problem now is that items duplicate locally because of the already implemented  addnewitem to list code, im taking a break now because im about to die. so far so good tho :)