rerouted items to sort inside the user
for some reason, i forgot to make the item database specific to each user and had to go around rerouting everything and got confused a little. turns out i was actually really close to the answer and i just had to add a another param when searching for the items and then displaying them. i guess the more you fail the more you learn but ideally i did not want to spend as much time as i did on this cuz it was literally just 1 freaking parameter (search for the specific id in database) but its whateeever ig. so far so good tho :)