added item confirm remove
when the item count is about to become zero, it shows a popup on the screen that wants to confirm if you want to delete the item or not. this will then have more options where you can move it to a specific shopping list in your grocery list. so far so good :)