fixed styling for new and login
felt a little tired tday cuz i was working on summatives and everything so i decided to spend some time doing something more lightweight but still having an impact on the project. now this stuff should be ready for the actual implementation of the store feature. so far so good :)