started store implementation
added a new screen that helps add a new store. it works pretty well right now but i also need to format the edit button for the grocery list to change the store and edit the popup to move the item to the grocery list. also realized that i dont need - items in grocery so i will change that to put a button there that can move it to the pantry if its bought. so far so good :)