updatable, displaying stores
stores now have a function to be displayed on the grocery list and can be edited in the edit menu. I also fixed the example item screen but i dont like how the code looks right now with all of the different Views. Next thing to work on is to make the popup when the item is moving from pantry to grocery and if so to what list. so far so good :)