itempopup moves items to specific stores
now when you remove a single item in the pantry list and it is about to be deleted, you have the option to move it to a store in the grocery list to keep track of items that you need to shop for easilly. there are still some other things i need to do like changing the function for the grocery popup to just delete the item or move it to pantry, adding amounts, and a path to make a new store if there isnt one already on the pantry move popup. so far so good :)