Barcode Scanning
somehow managed to get this running in 1 hour? still not fully finished yet but i connected it to openfoodfacts and it returns different categories and the item name itself. also returns generic item name which can be usefull. next thing that this will be able to do is to return some buttons that will be these identifiers and when clicked, it will forward you to a new item page where you can fill out the rest of the details. so far so good :)