continuing barcode scanning
working more on the barcode scanning, trying to set the item name properly and right now it has a list of items that show when clicked, but that itself was a bit confusing to setup and then i have to style it again. other than that, i need to make it route to the new page with the selected itemname as context which will be a little hard or easy depending on how my code is XD . really happy with how this is going though, i did not think i would be able to do this tday. so far so good :)