Finished barcode scanning
got the routing to work and got the name to transfer over to the new screen. also styled the popup to look a little better too. also formatted the data to be organized and split from commas because the categories sometimes have a lot of information in 1 section that needs to be split up. really happy with how its going. so far so good :)