added store organizer to list
you can now organize your items by the store in the list, a feature that i forgot to implement. I also added a shortcut to the scan item page from the homepage. you are also able to add a new list on the list tab itself. next thing that I want to do is to add the speech add to add an item by talking to your phone. so far so good :)