trying to get speech recognition setup
started working on this with expo speech recognition but couldnt get that to work, so i decided to switch to the google sst api but now i have to pay for that.  I will have to find another way to transcribe the text, but for now i have a screen setup with a button that records the audio. so far so good :)