rewriting auth code for android
spend a lot of time trying to figure out why android wasnt getting authorized after startup, then went into some stack overflow and github issues about firebase integration for android with this one specific function not working properly. after a lot of tinkering and fixing, i was able to get it to work by redirecting it to the signup page as usual and then checking if there is already an authorization and then pushing it to the homepage, which actually does the same thing and works on android. also did style fixes for android and am going to do more style fixes to make the app look better for all platforms. this was the first big problem i had with this app lol, so far so good :)