Fixed styling for web
Spent some time fixing all of the styling for the web devices. i spent a good amount of time trying to get the edit button to style properly and then decided to swap it with an icon instead which works great across all devices. I also changed the button component to look a little better and have more spacing for each button in the app. Also did more styling for the signup and login pages too. looks really nice on mobile and desktop web. so far so good :)