got local storage working again
decided to move back to local storage because it makes more sense to have the user save their data locally and not have to worry about login safety. also realized that the way i was storing user data was not safe and could easily be compromised, while async storage is reliable and safe by connecting items to keys for retreival. so far so good :)