Duplicate item info
added a feature that tells you that if you are in your grocery and you are adding an item that is the same item as it is in your pantry, you get a header that tells you that you have the same item in the pantry, do you want to continue? also added a function to merge duplicate items in the same list.