started similar item searching
started working on a new feature that uses an algorithm to search through the item and the list that the item is in to see if there are any other items that are closely related to the name of the selected item. then the user gets prompted to add as a new item or to add it to an item already on the list. then the user gets to see all of the items that are related to the selected item and pick one to add to. also deleted the save grocery list thing because it was redundant to the project. this simmilar item list looks pretty cool tho. so far so good :)