reactive camera, almost finished simmilar items
the camera can now turn off when the barcode is scanned to prevent double scanning. i am almost done with the simmilar item screen/checking but the only problem right now is that when the user is coming back from the screen and back to the scan screen, the camera wont turn on again. probably some logic error but im too tired to figure that out rn. I really like the design of the simmilar item.tsx and scan.tsx right now, and thanks to my family's opinions, the app is becoming better and better. depending on how much the simmilar item feature detects items, i will have to increase or decrease the threshold on what a similar item is. other than those things though, so far so good :)