new popup for setting sorted store
after a lot of different tries with different modal types, i decided this would be the best bet and decided to go with a modal that opens when the button is pressed and returns the selected store. This looks great on all platforms and works for different sizes of lists. so far so good :)