Modified button, ui fixes, fixed delete store, etc
Spent a good amount of time fixing the delete store button (if you look at the last devlog you can see that it is broken) and making the routing more cleaner and viewable. I also added the similar item check for manually added items. The home button now routes to the scan page properly by going to your list tab first. All of the buttons (including the ones in the list) use touchable opacity which makes 10x more sense now after using, its clear, easy, and gets the job done.  also removed that unnecessary and not usable use effect in the context for deleting the store. Did a lot of fixes that made the overall app look better, so far so good :)