creating mergable items
started implementing the mergable items feature. converted the edit button to a reactive icon that changes based on if you are trying to merge the items. you can select multiple items and then when done it forwards you to a placeholder screen where the user gets prompted to either put them under a new name, add them together under one unit and more things that i will have to think about when making it. so far so good :)