removed android folder, fixed merge into new name
spent a lot of time fixing the merge into new name option, turns out that i was passing the listtype wrongly into new.tsx, i didnt convert it to grocery/pantry and it stayed as first/second. this in turn caused the add new item function to stop working, so i fixed the way i passed the listtype as a param and now everything works as it should. so far so good :)