fixed delete and -1 item
the code that these features were using depended on the old firebase system and that had a listtype requirement which is no longer needed, fixed it by making the items  be deleted and then added to the other list with the same params and everything and the new selected listtype. still working on the merge items, but wanted to fix this bug first. so far so good :)