share list function
implemented a new function using rn-share and copy/paste, i formatted the lists to turn into a text format with only the count and amount and item name. this makes it so that you can share lists easily to your family/friends even if they dont have the app. so far so good ;)