ios build, dark mode, styling
got the apple dev certificate for hack club org and am going to upload the app to the app store to start getting feedback and review. i also made a dark mode and am still working on it to be throughout the app.  I removed all instances of firebase for now because i want this build to be completely local, now i will start to branch off and work on an account sharing/auth method, still want to have this there as a proper build tho. so far so good :)