fixed/finished merge items
after a lot of debugging and searching, i was able to implement the last feature (i think i forgot about it) to merge the items into a new name. Still want to do one more thing where i can have the user merge the units tho so that might take some time. so far so good :)