ocr server connected to app!
cant belive i got this to work! the entire api for scanning is done via pytessarect an open source OCR library that will return the name of the item. I've also started to style everything but will also start to add buttons and everything else needed to make it functional and ready. (also if hack club nest could work that would be great for hosting!)